# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Farm Management API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "security": [
        {
            "apiKey": []
        }
    ],
    "paths": {
...
        "/subscription/entitlements": {
            "get": {
                "operationId": "v1.subscription.entitlements",
                "description": "What the farm's plan allows right now: boolean `features` and numeric `limits` (`unlimited: true` with\n`limit: null` means no cap). Use this to enable/disable UI; the API enforces the same rules and answers\ndenied actions with `feature_not_available`, `subscription_inactive` or `plan_limit_reached`.\nAny farm member may read this (`farm.view`); it does not grant any RBAC permission.\n\n**Requires farm permission:** `farm.view`.",
                "summary": "Current entitlements",
                "tags": [
                    "Subscription"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "plan": {
                                                    "type": [
                                                        "object",
                                                        "null"
                                                    ],
                                                    "properties": {
                                                        "id": {
                                                            "type": "string"
                                                        },
                                                        "slug": {
                                                            "type": "string"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": [
                                                        "id",
                                                        "slug",
                                                        "name"
                                                    ]
                                                },
                                                "subscription_inactive": {
                                                    "type": "boolean"
                                                },
                                                "features": {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "limits": {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "object",
                                                        "properties": {
                                                            "limit": {
                                                                "type": [
                                                                    "integer",
                                                                    "null"
                                                                ]
                                                            },
                                                            "unlimited": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "required": [
                                                            "limit",
                                                            "unlimited"
                                                        ]
                                                    }
                                                }
                                            },
                                            "required": [
                                                "plan",
                                                "subscription_inactive",
                                                "features",
                                                "limits"
                                            ]
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: account state does not allow this request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable summary."
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Machine-readable error code.",
                                            "enum": [
                                                "account_suspended",
                                                "email_verification_required",
                                                "onboarding_required",
                                                "no_active_farm",
                                                "farm_access_denied",
                                                "forbidden"
                                            ]
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Echo of the X-Request-Id header, for support."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional structured context. Entitlement errors (feature_not_available, subscription_inactive, plan_limit_reached) carry entitlement_key and, for limits, limit, usage and remaining."
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "code",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Sanctum session cookie set by /auth/login, /auth/register or /auth/google. Send requests with credentials and, for POST/PUT/PATCH/DELETE, the X-XSRF-TOKEN header (see docs/api/README.md).",
                "in": "cookie",
                "name": "farm_management_api_session"
            }
        },
...
    }
}